Use waitForGethConnectivity, rearrange sagas
