+++ packages/mobile/src/account/actions.ts
@@ -23,6 +23,7 @@ export enum Actions {
   DISMISS_INVITE_FRIENDS = 'ACCOUNT/DISMISS_INVITE_FRIENDS',
   DISMISS_GET_VERIFIED = 'ACCOUNT/DISMISS_GET_VERIFIED',
   SET_USER_CONTACT_DETAILS = 'ACCOUNT/SET_USER_CONTACT_DETAILS',
+  SET_PROMPT_FORNO = 'GETH/SET_PROMPT_FORNO',
 }
 
 export interface SetNameAction {
@@ -107,6 +108,11 @@ export interface SetContactDetailsAction {
   thumbnailPath: string | null
 }
 
+interface SetPromptFornoAction {
+  type: Actions.SET_PROMPT_FORNO
+  promptIfNeeded: boolean
+}
+
 export type ActionTypes =
   | SetNameAction
   | SetPhoneNumberAction
@@ -126,6 +132,7 @@ export type ActionTypes =
   | UpdateIncomingPaymentRequestsAction
   | UpdateOutgoingPaymentRequestsAction
   | SetContactDetailsAction
+  | SetPromptFornoAction
 
 export function setName(name: string): SetNameAction {
   return {
@@ -210,6 +217,11 @@ export const dismissInviteFriends = (): DismissInviteFriendsAction => ({
 
 export const dismissGetVerified = (): DismissGetVerifiedAction => ({
   type: Actions.DISMISS_GET_VERIFIED,
+})
+
+export const setPromptForno = (promptIfNeeded: boolean): SetPromptFornoAction => ({
+  type: Actions.SET_PROMPT_FORNO,
+  promptIfNeeded,
 })
 
 export const setUserContactDetails = (
