+++ packages/mobile/src/geth/actions.ts
@@ -4,7 +4,6 @@ export enum Actions {
   SET_INIT_STATE = 'GETH/SET_INIT_STATE',
   SET_GETH_CONNECTED = 'GETH/SET_GETH_CONNECTED',
   CANCEL_GETH_SAGA = 'GETH/CANCEL_GETH_SAGA',
-  SET_PROMPT_FORNO = 'GETH/SET_PROMPT_FORNO',
 }
 
 interface SetInitStateAction {
@@ -21,16 +20,6 @@ export const cancelGethSaga = () => ({
   type: Actions.CANCEL_GETH_SAGA,
 })
 
-interface SetPromptFornoAction {
-  type: Actions.SET_PROMPT_FORNO
-  promptIfNeeded: boolean
-}
-
-export const setPromptForno = (promptIfNeeded: boolean): SetPromptFornoAction => ({
-  type: Actions.SET_PROMPT_FORNO,
-  promptIfNeeded,
-})
-
 interface SetGethConnectedAction {
   type: Actions.SET_GETH_CONNECTED
   connected: boolean
@@ -41,4 +30,4 @@ export const setGethConnected = (connected: boolean): SetGethConnectedAction => 
   connected,
 })
 
-export type ActionTypes = SetInitStateAction | SetGethConnectedAction | SetPromptFornoAction
+export type ActionTypes = SetInitStateAction | SetGethConnectedAction
