+++ packages/mobile/src/geth/reducer.ts
@@ -11,28 +11,28 @@ export enum InitializationState {
 export interface State {
   initialized: InitializationState
   connected: boolean
-  promptFornoIfNeeded: boolean
+  gethStartedThisSession: boolean
 }
 
 const initialState: State = {
   initialized: InitializationState.NOT_YET_INITIALIZED,
   connected: false,
-  promptFornoIfNeeded: false,
+  gethStartedThisSession: false,
 }
 
 export function gethReducer(state: State = initialState, action: ActionTypes) {
   switch (action.type) {
     case Actions.SET_INIT_STATE:
-      return { ...state, initialized: action.state }
+      return {
+        ...state,
+        initialized: action.state,
+        gethStartedThisSession:
+          action.state === InitializationState.INITIALIZED ? true : state.gethStartedThisSession, // Once geth initialized, it has been started this session
+      }
     case Actions.SET_GETH_CONNECTED:
       return {
         ...state,
         connected: action.connected,
-      }
-    case Actions.SET_PROMPT_FORNO:
-      return {
-        ...state,
-        promptFornoIfNeeded: action.promptIfNeeded,
       }
     default:
       return state
