+++ packages/mobile/test/schemas.ts
@@ -236,13 +236,17 @@ export const v5Schema = {
     ...v4Schema.localCurrency,
     exchangeRate: '1.33',
   },
-  geth: {
-    ...v4Schema.geth,
-    promptFornoIfNeeded: false,
-  },
   web3: {
     ...v4Schema.web3,
     fornoMode: false,
+  },
+  geth: {
+    ...v4Schema.geth,
+    gethStartedThisSession: false,
+  },
+  account: {
+    ...v4Schema.account,
+    promptFornoIfNeeded: false,
   },
 }
